cleaned up the class structure for initializing speeches
